refactor: replace .flatMap(f => f) anti-pattern with .flatten (#1181)
* refactor: replace .flatMap(f => f) anti-pattern with .flatten

.flatMap(f => f) is a verbose identity flatMap — the idiomatic Scala
equivalent is simply .flatten. Applied mechanically across 31 files
(115 occurrences) to improve readability with no behaviour change.

Affected modules: taxonomy-actors, FrameworkManager, FrameworkActor,
CategoryInstanceActor, ObjectCategoryDefinitionActor, ContentActor,
HierarchyManager, UpdateHierarchyManager, DataNode, DefinitionNode,
RelationValidator, VersioningNode, VersionKeyValidator,
FrameworkValidator, CopyManager, FlagManager, AcceptFlagManager,
AssetCopyManager, DiscardManager, DIALManager, PublishManager,
ReviewManager, UploadManager, ItemSetActor, QuestionActor,
QuestionSetActor, assessment CopyManager, parseq Task,
CollectionMimeTypeMgrImpl.


* fix: replace non-existent TelemetryManager.debug with TelemetryManager.log to resolve build errors

* Replace JavaConverters with jdk.CollectionConverters

Switch scala.collection.JavaConverters to scala.jdk.CollectionConverters and simplify conversions by replacing JavaConverters.seqAsJavaListConverter(...).asJava with .toList.asJava in QuestionActor and QuestionSetActor. This removes deprecated API usage (Scala 2.13+ compatibility) and keeps runtime behavior unchanged.

---------
