+++ content-api/content-actors/src/main/scala/org/sunbird/content/publish/mgr/PublishManager.scala
@@ -43,7 +43,7 @@ object PublishManager {
 			response.put(ContentConstants.NODE_ID, node.getIdentifier)
 
 			Future(response)
-		}).flatMap(f => f)
+		}).flatten
 	}
 
 	@throws[Exception]
