+++ content-api/content-actors/src/main/scala/org/sunbird/content/review/mgr/ReviewManager.scala
@@ -24,7 +24,7 @@ object ReviewManager {
 			DataNode.update(updateReq).map(node => {
 				ResponseHandler.OK.putAll(Map("identifier" -> node.getIdentifier.replace(".img", ""), "versionKey" -> node.getMetadata.get("versionKey")).asJava)
 			})
-		}).flatMap(f => f)
+		}).flatten
 	}
 }
