+++ ontology-engine/graph-engine_2.13/src/main/scala/org/sunbird/graph/schema/DefinitionNode.scala
@@ -130,7 +130,7 @@ object DefinitionNode {
         node
       })
 
-    }).flatMap(f => f)
+    }).flatten
   }
 
   def postProcessor(request: Request, node: Node)(implicit ec: ExecutionContext, oec: OntologyEngineContext): Node = {
